+++ src/action.sh
@@ -15,7 +15,24 @@ if jq -e '[.runs[].results[]] | length == 0' < .checkton.sarif >/dev/null; then
     echo "✅ No ShellCheck warnings, congrats! \\o/"
 else
     echo "❌ Found ShellCheck warnings! /o\\"
-    csgrep --embed-context 4 < .checkton.sarif
+
+    if [[ "${GITHUB_ACTIONS:-false}" == true ]]; then
+        # csgrep colored output doesn't look very good in GH actions, use colors only for sarif-fmt
+        sarif_fmt_color=always
+    else
+        sarif_fmt_color=auto
+    fi
+
+    case "${CHECKTON_DISPLAY_STYLE:-sarif-fmt}" in
+        csgrep)
+            csgrep --embed 4 < .checkton.sarif ;;
+        sarif-fmt)
+            sarif-fmt --color="$sarif_fmt_color" < .checkton.sarif ;;
+        *)
+            echo "Unknown display style: $CHECKTON_DISPLAY_STYLE" >&2
+            exit 1
+            ;;
+    esac
 
     if [[ "${CHECKTON_FAIL_ON_FINDINGS:-}" == "true" ]]; then
         exitcode=1